+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file, following t
 ## [Unreleased]
 
 - Selections support `type_symbol`
+- Complex superposition extension allows RNA superposition
 
 ## [v3.6.0] - 2025-08-12
 - Option `hideCanvasControls` accepts "screenshot" value